Create separate pipelines for each add-style invocation
This means in practice that scss will be processed through a separate
pipeline from css. Before this change they would share pipeline,
practically resulting in that all CSS in our project would start being
processed through Node Sass.

The hook API stays the same and because of this no changes are needed in
any of the other style plugins.